requirements.txt
langchain-groq
python-dotenv
langchain
pdfplumber
python-docx
pydantic
streamlit
langchain-core